extract gql queries
